// app-ui/src/app/backend/model/analysisMetrics.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AnalysisMetrics { 
    numEvents?: number;
    numCases?: number;
    numActivities?: number;
    numVariants?: number;
    numResources?: number;
    avgTime?: number;
    medianTime?: number;
    minTime?: number;
    maxTime?: number;
    avgActivities?: number;
    minActivities?: number;
    maxActivities?: number;
    avgResourcesPerCases?: number;
    minResourcesPerCases?: number;
    maxResourcesPerCases?: number;
    minTimestamp?: string;
    maxTimestamp?: string;
}